requirements.txt
# LLM Framework
langchain
langchain-community
langchain-google-genai
google-generativeai

# Environment Variables
python-dotenv

# Vector Databases
chromadb

# Data Processing
pandas
numpy

# Web UI
streamlit

